+++ src/recipe.h
@@ -97,6 +97,13 @@ namespace forge
     std::string suffix;
   };
 
+  struct ReleaseReadme
+  {
+    std::filesystem::path linux;
+    std::filesystem::path macos;
+    std::filesystem::path windows;
+  };
+
   struct Recipe
   {
     std::string name;
@@ -126,6 +133,7 @@ namespace forge
     std::vector<std::filesystem::path> release_files;
     std::optional<std::string> release_bundle_name;
     std::vector<ReleaseVariant> release_variants;
+    ReleaseReadme release_readme;
     std::optional<std::string> release_notes_build_number_format;
     std::filesystem::path version_header_path;
     std::string version_header_prefix;